Moving bbox-related functionality to bbox module
