+++ kornia/geometry/__init__.py
@@ -7,3 +7,4 @@ from kornia.geometry.homography import *
 from kornia.geometry.linalg import *
 from kornia.geometry.subpix import *
 from kornia.geometry.transform import *
+from kornia.geometry.bbox import *
