+++ kornia/geometry/linalg.py
@@ -9,7 +9,6 @@ __all__ = [
     "relative_transformation",
     "inverse_transformation",
     "transform_points",
-    "transform_boxes",
     "perspective_transform_lafs",
 ]
 
@@ -203,45 +202,6 @@ def transform_points(trans_01: torch.Tensor, points_1: torch.Tensor) -> torch.Te
     shape_inp[-1] = points_0.shape[-1]
     points_0 = points_0.reshape(shape_inp)
     return points_0
-
-
-def transform_boxes(trans_mat: torch.Tensor, boxes: torch.Tensor, mode: str = "xyxy") -> torch.Tensor:
-    r"""Function that applies a transformation matrix to a box or batch of boxes. Boxes must
-    be a tensor of the shape (N, 4) or a batch of boxes (B, N, 4) and trans_mat must be a (3, 3)
-    transformation matrix or a batch of transformation matrices (B, 3, 3)
-
-    Args:
-        trans_mat (torch.Tensor): The transformation matrix to be applied
-        boxes (torch.Tensor): The boxes to be transformed
-        mode (str): The format in which the boxes are provided. If set to 'xyxy' the boxes
-                    are assumed to be in the format (xmin, ymin, xmax, ymax). If set to 'xywh'
-                    the boxes are assumed to be in the format (xmin, ymin, width, height).
-                    Default: 'xyxy'
-    Returns:
-        torch.Tensor: The set of transformed points in the specified mode
-
-
-    """
-
-    if not isinstance(mode, str):
-        raise TypeError(f"Mode must be a string. Got {type(mode)}")
-
-    if mode not in ("xyxy", "xywh"):
-        raise ValueError(f"Mode must be one of 'xyxy', 'xywh'. Got {mode}")
-
-    # convert boxes to format xyxy
-    if mode == "xywh":
-        boxes[..., -2] = boxes[..., 0] + boxes[..., -2]  # x + w
-        boxes[..., -1] = boxes[..., 1] + boxes[..., -1]  # y + h
-
-    transformed_boxes: torch.Tensor = kornia.transform_points(trans_mat, boxes.view(boxes.shape[0], -1, 2))
-    transformed_boxes = transformed_boxes.view_as(boxes)
-
-    if mode == 'xywh':
-        transformed_boxes[..., 2] = transformed_boxes[..., 2] - transformed_boxes[..., 0]
-        transformed_boxes[..., 3] = transformed_boxes[..., 3] - transformed_boxes[..., 1]
-
-    return transformed_boxes
 
 
 def perspective_transform_lafs(trans_01: torch.Tensor, lafs_1: torch.Tensor) -> torch.Tensor:
