+++ kornia/geometry/transform/crop/crop2d.py
@@ -3,16 +3,13 @@ from typing import Optional, Tuple
 import torch
 
 from kornia.geometry.transform.imgwarp import get_perspective_transform, warp_affine
+from kornia.geometry.bbox import validate_bbox_2d, infer_bbox_shape_2d
 
 __all__ = [
     "crop_and_resize",
     "crop_by_boxes",
     "crop_by_transform_mat",
     "center_crop",
-    "bbox_to_mask",
-    "infer_box_shape",
-    "validate_bboxes",
-    "bbox_generator",
 ]
 
 
@@ -216,8 +213,8 @@ def crop_by_boxes(
         RuntimeError: solve_cpu: For batch 0: U(2,2) is zero, singular U.
     """
     # TODO: improve this since might slow down the function
-    validate_bboxes(src_box)
-    validate_bboxes(dst_box)
+    validate_bbox_2d(src_box)
+    validate_bbox_2d(dst_box)
 
     assert len(tensor.shape) == 4, f"Only tensor with shape (B, C, H, W) supported. Got {tensor.shape}."
 
@@ -225,7 +222,7 @@ def crop_by_boxes(
     # Note: Tensor.dtype must be float. "solve_cpu" not implemented for 'Long'
     dst_trans_src: torch.Tensor = get_perspective_transform(src_box.to(tensor), dst_box.to(tensor))
 
-    bbox: Tuple[torch.Tensor, torch.Tensor] = infer_box_shape(dst_box)
+    bbox: Tuple[torch.Tensor, torch.Tensor] = infer_bbox_shape_2d(dst_box)
     assert (bbox[0] == bbox[0][0]).all() and (bbox[1] == bbox[1][0]).all(), (
         f"Cropping height, width and depth must be exact same in a batch. " f"Got height {bbox[0]} and width {bbox[1]}."
     )
