+++ kornia/geometry/transform/crop/crop3d.py
@@ -3,16 +3,13 @@ from typing import Optional, Tuple
 import torch
 
 from kornia.geometry.transform.projwarp import get_perspective_transform3d, warp_affine3d
+from kornia.geometry.bbox import validate_bbox_3d, infer_bbox_shape_3d
 
 __all__ = [
     "crop_and_resize3d",
     "crop_by_boxes3d",
     "crop_by_transform_mat3d",
     "center_crop3d",
-    "bbox_to_mask3d",
-    "infer_box_shape3d",
-    "validate_bboxes3d",
-    "bbox_generator3d",
 ]
 
 
@@ -307,8 +304,8 @@ def crop_by_boxes3d(
                    [45., 46., 47.]]]]])
 
     """
-    validate_bboxes3d(src_box)
-    validate_bboxes3d(dst_box)
+    validate_bbox_3d(src_box)
+    validate_bbox_3d(dst_box)
 
     assert len(tensor.shape) == 5, f"Only tensor with shape (B, C, D, H, W) supported. Got {tensor.shape}."
 
@@ -318,7 +315,7 @@ def crop_by_boxes3d(
     # simulate broadcasting
     dst_trans_src = dst_trans_src.expand(tensor.shape[0], -1, -1).type_as(tensor)
 
-    bbox = infer_box_shape3d(dst_box)
+    bbox = infer_bbox_shape_3d(dst_box)
     assert (bbox[0] == bbox[0][0]).all() and (bbox[1] == bbox[1][0]).all() and (bbox[2] == bbox[2][0]).all(), (
         "Cropping height, width and depth must be exact same in a batch."
         f"Got height {bbox[0]}, width {bbox[1]} and depth {bbox[2]}."
@@ -366,260 +363,3 @@ def crop_by_transform_mat3d(
     )
 
     return patches
-
-
-def infer_box_shape3d(boxes: torch.Tensor) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
-    r"""Auto-infer the output sizes for the given 3D bounding boxes.
-
-    Args:
-        boxes (torch.Tensor): a tensor containing the coordinates of the bounding boxes to be extracted.
-            The tensor must have the shape of Bx8x3, where each box is defined in the following (clockwise)
-            order: front-top-left, front-top-right, front-bottom-right, front-bottom-left, back-top-left,
-            back-top-right, back-bottom-right, back-bottom-left. The coordinates must be in the x, y, z order.
-
-    Returns:
-        Tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
-        - Bounding box depths, shape of :math:`(B,)`.
-        - Bounding box heights, shape of :math:`(B,)`.
-        - Bounding box widths, shape of :math:`(B,)`.
-
-    Example:
-        >>> boxes = torch.tensor([[[ 0,  1,  2],
-        ...         [10,  1,  2],
-        ...         [10, 21,  2],
-        ...         [ 0, 21,  2],
-        ...         [ 0,  1, 32],
-        ...         [10,  1, 32],
-        ...         [10, 21, 32],
-        ...         [ 0, 21, 32]],
-        ...        [[ 3,  4,  5],
-        ...         [43,  4,  5],
-        ...         [43, 54,  5],
-        ...         [ 3, 54,  5],
-        ...         [ 3,  4, 65],
-        ...         [43,  4, 65],
-        ...         [43, 54, 65],
-        ...         [ 3, 54, 65]]]) # 2x8x3
-        >>> infer_box_shape3d(boxes)
-        (tensor([31, 61]), tensor([21, 51]), tensor([11, 41]))
-    """
-    validate_bboxes3d(boxes)
-
-    left = torch.index_select(boxes, 1, torch.tensor([1, 2, 5, 6], device=boxes.device, dtype=torch.long))[:, :, 0]
-    right = torch.index_select(boxes, 1, torch.tensor([0, 3, 4, 7], device=boxes.device, dtype=torch.long))[:, :, 0]
-    widths = (left - right + 1)[:, 0]
-
-    bot = torch.index_select(boxes, 1, torch.tensor([2, 3, 6, 7], device=boxes.device, dtype=torch.long))[:, :, 1]
-    upper = torch.index_select(boxes, 1, torch.tensor([0, 1, 4, 5], device=boxes.device, dtype=torch.long))[:, :, 1]
-    heights = (bot - upper + 1)[:, 0]
-
-    depths = (boxes[:, 4:, 2] - boxes[:, :4, 2] + 1)[:, 0]
-    return (depths, heights, widths)
-
-
-def validate_bboxes3d(boxes: torch.Tensor) -> None:
-    """Validate if a 3D bounding box usable or not.
-
-    This function checks if the boxes are cube or not.
-
-    Args:
-        boxes (torch.Tensor): a tensor containing the coordinates of the bounding boxes to be extracted.
-            The tensor must have the shape of Bx8x3, where each box is defined in the following (clockwise)
-            order: front-top-left, front-top-right, front-bottom-right, front-bottom-left, back-top-left,
-            back-top-right, back-bottom-right, back-bottom-left. The coordinates must be in the x, y, z order.
-    """
-    assert len(boxes.shape) == 3 and boxes.shape[1:] == torch.Size(
-        [8, 3]
-    ), f"Box shape must be (B, 8, 3). Got {boxes.shape}."
-
-    left = torch.index_select(boxes, 1, torch.tensor([1, 2, 5, 6], device=boxes.device, dtype=torch.long))[:, :, 0]
-    right = torch.index_select(boxes, 1, torch.tensor([0, 3, 4, 7], device=boxes.device, dtype=torch.long))[:, :, 0]
-    widths = left - right + 1
-    assert torch.allclose(
-        widths.permute(1, 0), widths[:, 0]
-    ), f"Boxes must have be cube, while get different widths {widths}."
-
-    bot = torch.index_select(boxes, 1, torch.tensor([2, 3, 6, 7], device=boxes.device, dtype=torch.long))[:, :, 1]
-    upper = torch.index_select(boxes, 1, torch.tensor([0, 1, 4, 5], device=boxes.device, dtype=torch.long))[:, :, 1]
-    heights = bot - upper + 1
-    assert torch.allclose(
-        heights.permute(1, 0), heights[:, 0]
-    ), f"Boxes must have be cube, while get different heights {heights}."
-
-    depths = boxes[:, 4:, 2] - boxes[:, :4, 2] + 1
-    assert torch.allclose(
-        depths.permute(1, 0), depths[:, 0]
-    ), f"Boxes must have be cube, while get different depths {depths}."
-
-
-def bbox_to_mask3d(boxes: torch.Tensor, size: Tuple[int, int, int]) -> torch.Tensor:
-    """Convert 3D bounding boxes to masks. Covered area is 1. and the remaining is 0.
-
-    Args:
-        boxes (torch.Tensor): a tensor containing the coordinates of the bounding boxes to be extracted.
-            The tensor must have the shape of Bx8x3, where each box is defined in the following (clockwise)
-            order: front-top-left, front-top-right, front-bottom-right, front-bottom-left, back-top-left,
-            back-top-right, back-bottom-right, back-bottom-left. The coordinates must be in the x, y, z order.
-        size (Tuple[int, int, int]): depth, height and width of the masked image.
-
-    Returns:
-        torch.Tensor: the output mask tensor.
-
-    Examples:
-        >>> boxes = torch.tensor([[
-        ...     [1., 1., 1.],
-        ...     [2., 1., 1.],
-        ...     [2., 2., 1.],
-        ...     [1., 2., 1.],
-        ...     [1., 1., 2.],
-        ...     [2., 1., 2.],
-        ...     [2., 2., 2.],
-        ...     [1., 2., 2.],
-        ... ]])  # 1x8x3
-        >>> bbox_to_mask3d(boxes, (4, 5, 5))
-        tensor([[[[[0., 0., 0., 0., 0.],
-                   [0., 0., 0., 0., 0.],
-                   [0., 0., 0., 0., 0.],
-                   [0., 0., 0., 0., 0.],
-                   [0., 0., 0., 0., 0.]],
-        <BLANKLINE>
-                  [[0., 0., 0., 0., 0.],
-                   [0., 1., 1., 0., 0.],
-                   [0., 1., 1., 0., 0.],
-                   [0., 0., 0., 0., 0.],
-                   [0., 0., 0., 0., 0.]],
-        <BLANKLINE>
-                  [[0., 0., 0., 0., 0.],
-                   [0., 1., 1., 0., 0.],
-                   [0., 1., 1., 0., 0.],
-                   [0., 0., 0., 0., 0.],
-                   [0., 0., 0., 0., 0.]],
-        <BLANKLINE>
-                  [[0., 0., 0., 0., 0.],
-                   [0., 0., 0., 0., 0.],
-                   [0., 0., 0., 0., 0.],
-                   [0., 0., 0., 0., 0.],
-                   [0., 0., 0., 0., 0.]]]]])
-    """
-    validate_bboxes3d(boxes)
-    mask = torch.zeros((len(boxes), *size))
-
-    mask_out = []
-    # TODO: Looking for a vectorized way
-    for m, box in zip(mask, boxes):
-        m = m.index_fill(
-            0,
-            torch.arange(box[0, 2].item(), box[4, 2].item() + 1, device=box.device, dtype=torch.long),
-            torch.tensor(1, device=box.device, dtype=box.dtype),
-        )
-        m = m.index_fill(
-            1,
-            torch.arange(box[1, 1].item(), box[2, 1].item() + 1, device=box.device, dtype=torch.long),
-            torch.tensor(1, device=box.device, dtype=box.dtype),
-        )
-        m = m.index_fill(
-            2,
-            torch.arange(box[0, 0].item(), box[1, 0].item() + 1, device=box.device, dtype=torch.long),
-            torch.tensor(1, device=box.device, dtype=box.dtype),
-        )
-        m = m.unsqueeze(dim=0)
-        m_out = torch.ones_like(m)
-        m_out = m_out * (m == 1).all(dim=2, keepdim=True).all(dim=1, keepdim=True)
-        m_out = m_out * (m == 1).all(dim=3, keepdim=True).all(dim=1, keepdim=True)
-        m_out = m_out * (m == 1).all(dim=2, keepdim=True).all(dim=3, keepdim=True)
-        mask_out.append(m_out)
-
-    return torch.stack(mask_out, dim=0).float()
-
-
-def bbox_generator3d(
-    x_start: torch.Tensor,
-    y_start: torch.Tensor,
-    z_start: torch.Tensor,
-    width: torch.Tensor,
-    height: torch.Tensor,
-    depth: torch.Tensor,
-) -> torch.Tensor:
-    """Generate 3D bounding boxes according to the provided start coords, width, height and depth.
-
-    Args:
-        x_start (torch.Tensor): a tensor containing the x coordinates of the bounding boxes to be extracted.
-            Shape must be a scalar tensor or :math:`(B,)`.
-        y_start (torch.Tensor): a tensor containing the y coordinates of the bounding boxes to be extracted.
-            Shape must be a scalar tensor or :math:`(B,)`.
-        z_start (torch.Tensor): a tensor containing the z coordinates of the bounding boxes to be extracted.
-            Shape must be a scalar tensor or :math:`(B,)`.
-        width (torch.Tensor): widths of the masked image.
-            Shape must be a scalar tensor or :math:`(B,)`.
-        height (torch.Tensor): heights of the masked image.
-            Shape must be a scalar tensor or :math:`(B,)`.
-        depth (torch.Tensor): depths of the masked image.
-            Shape must be a scalar tensor or :math:`(B,)`.
-
-    Returns:
-        torch.Tensor: the 3d bounding box tensor :math:`(B, 8, 3)`.
-
-    Examples:
-        >>> x_start = torch.tensor([0, 3])
-        >>> y_start = torch.tensor([1, 4])
-        >>> z_start = torch.tensor([2, 5])
-        >>> width = torch.tensor([10, 40])
-        >>> height = torch.tensor([20, 50])
-        >>> depth = torch.tensor([30, 60])
-        >>> bbox_generator3d(x_start, y_start, z_start, width, height, depth)
-        tensor([[[ 0,  1,  2],
-                 [10,  1,  2],
-                 [10, 21,  2],
-                 [ 0, 21,  2],
-                 [ 0,  1, 32],
-                 [10,  1, 32],
-                 [10, 21, 32],
-                 [ 0, 21, 32]],
-        <BLANKLINE>
-                [[ 3,  4,  5],
-                 [43,  4,  5],
-                 [43, 54,  5],
-                 [ 3, 54,  5],
-                 [ 3,  4, 65],
-                 [43,  4, 65],
-                 [43, 54, 65],
-                 [ 3, 54, 65]]])
-    """
-    assert x_start.shape == y_start.shape == z_start.shape and x_start.dim() in [
-        0,
-        1,
-    ], f"`x_start`, `y_start` and `z_start` must be a scalar or (B,). Got {x_start}, {y_start}, {z_start}."
-    assert width.shape == height.shape == depth.shape and width.dim() in [
-        0,
-        1,
-    ], f"`width`, `height` and `depth` must be a scalar or (B,). Got {width}, {height}, {depth}."
-    assert x_start.dtype == y_start.dtype == z_start.dtype == width.dtype == height.dtype == depth.dtype, (
-        "All tensors must be in the same dtype. "
-        f"Got `x_start`({x_start.dtype}), `y_start`({x_start.dtype}), `z_start`({x_start.dtype}), "
-        f"`width`({width.dtype}), `height`({height.dtype}) and `depth`({depth.dtype})."
-    )
-    assert x_start.device == y_start.device == z_start.device == width.device == height.device == depth.device, (
-        "All tensors must be in the same device. "
-        f"Got `x_start`({x_start.device}), `y_start`({x_start.device}), `z_start`({x_start.device}), "
-        f"`width`({width.device}), `height`({height.device}) and `depth`({depth.device})."
-    )
-
-    # front
-    bbox = torch.tensor(
-        [[[0, 0, 0], [0, 0, 0], [0, 0, 0], [0, 0, 0]]], device=x_start.device, dtype=x_start.dtype
-    ).repeat(len(x_start), 1, 1)
-
-    bbox[:, :, 0] += x_start.view(-1, 1)
-    bbox[:, :, 1] += y_start.view(-1, 1)
-    bbox[:, :, 2] += z_start.view(-1, 1)
-    bbox[:, 1, 0] += width
-    bbox[:, 2, 0] += width
-    bbox[:, 2, 1] += height
-    bbox[:, 3, 1] += height
-
-    # back
-    bbox_back = bbox.clone()
-    bbox_back[:, :, -1] += depth.unsqueeze(dim=1).repeat(1, 4)
-    bbox = torch.cat([bbox, bbox_back], dim=1)
-
-    return bbox
